Mao implementation (#30)
* Partially Mao theme implementation

Needs to be confirmed that the look it right.
several small changes/fixes to make it work in our omeka/seasons version (and not using several of the plugins mao was using)

- Hard sets the seasons theme to be autumn only (and removes other options from css so its easier to maintain like this)
- hard copied css values into sass (there was enough changes and small adjustments everywhere and the original sass was missing so it was quicker to just hard code all the variables in)
- Fixes the css for the Multilanguage locale switcher & word cloud plugins

in theme setting rememeber to:
- remove locale switcher from header in settings (moved to be hard coded into navbar)

* Initial work on style tweaks

A lot of hard coded style changes, but most of these are in styles.scss. These are mostly spacing things etc, but some small adjustments to layout as well.

* Small tweaks

* Significant work on heading and getting logo integrated

---------

Co-authored-by: Andrew Gardener <andrew.e.gardener@gmail.com>